add new entities

+++ src/main/java/org/ntg/spring/dto/CompanyReadDto.java
@@ -1,4 +1,4 @@
 package org.ntg.spring.dto;
 
-public record CompanyReadDto(Long id) {
+public record CompanyReadDto(Integer id) {
 }
